Add stubs for doubly linked list (and iterator) tests.
