+++ tests/lists/test_doubly_linked_list_iterator.py
@@ -1,0 +1,20 @@
+import unittest
+
+from aed_ds.lists.double_linked_list import DoublyLinkedList
+from aed_ds.exceptions import NoSuchElementException
+
+class TestSinglyLinkedListIterator(unittest.TestCase):
+    def setUp(self):
+        pass
+
+    # def test_has_next(self): pass
+
+    # def test_next(self): pass
+
+    # def test_rewind(self): pass
+
+    # def test_has_previous(self): pass
+    
+    # def test_previous(self): pass
+
+    # def test_full_forward(self): pass
